backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.5
python-jose==3.3.0
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.12.0
email-validator==2.1.0
supabase==2.3.0
boto3==1.34.2
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
reportlab==4.0.7
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
scipy==1.11.4
sqlalchemy==2.0.23
google-auth==2.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
